requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-groq
langchain-community
reportlab
streamlit
requests
